requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
Pillow>=8.0,<9.0
django-cors-headers>=3.10,<4.0
psycopg2-binary>=2.9,<3.0  # Si se utiliza PostgreSQL como base de datos
# Agrega otras dependencias según sea necesario para tu proyecto